+++ app/src/main/shared/components/DatePicker/DateRangePicker.tsx
@@ -1,0 +1,47 @@
+import * as React from "react";
+import {DatePicker} from "./DatePicker";
+import {CalendarIcon} from "./Calendar";
+
+interface DateRangePickerProps {
+    value: { start: Date, end: Date },
+    startDate: Date,
+    endDate: Date,
+    timeFromSelected: (date: Date) => void;
+    timeUntilSelected: (date: Date) => void;
+}
+
+export const DateRangePicker: React.StatelessComponent<DateRangePickerProps> = (props: DateRangePickerProps) => (
+    <div className="report-date-filters">
+        <div className="d-inline-block date-filter">
+            <div className="input-group">
+                <div className="input-group-prepend">
+                <span className="input-group-text">
+                    <CalendarIcon/>
+                </span>
+                </div>
+                <DatePicker
+                    onChange={props.timeFromSelected}
+                    value={props.value.start}
+                    fromMonth={props.startDate}
+                    toMonth={props.endDate}
+                />
+            </div>
+        </div>
+        <span> to </span>
+        <div className="d-inline-block date-filter">
+            <div className="input-group picker-on-right">
+                <div className="input-group-prepend">
+                <span className="input-group-text">
+                    <CalendarIcon/>
+                </span>
+                </div>
+                <DatePicker
+                    onChange={props.timeUntilSelected}
+                    value={props.value.end}
+                    fromMonth={props.startDate}
+                    toMonth={props.endDate}
+                />
+            </div>
+        </div>
+    </div>
+);
